+++ app/db/session.py
@@ -3,8 +3,8 @@ from dataclasses import dataclass, field
 from sqlalchemy import create_engine, Engine
 from sqlalchemy.orm import sessionmaker
 from typing import Any
-from app.utils.uitls import LoggerAdapter
-from app.utils.uitls import DATABASE_URL
+from app.utils.utils import LoggerAdapter
+from app.utils.utils import DATABASE_URL
 
 default_logger = logging.getLogger(__name__)
